Inline WC_Stripe_Settings_Controller into WC_Stripe
Moves the settings page admin hooks (script enqueue, options wrapper,
refund button hide, gateway list filter, OAuth AJAX) from the controller
class onto WC_Stripe, mirroring how this logic lived directly on
WC_Gateway_Stripe before it was extracted in #1713 (Aug 2021).

The controller is kept as a deprecation shim so external instantiations
do not break; each entry point delegates to WC_Stripe and emits a
_deprecated_function notice.

The PHPStan baseline is regenerated to reflect the relocation; no new
errors were silenced — only existing entries followed the methods.
